+++ R/rblimp_source.R
@@ -12,7 +12,8 @@
 #' Running `rblimp_source` will also run a check to see if Blimp is up to date.
 #' If Blimp is not up to date, it will prompt the user if it would like to update or not.
 #' This check will only be performed on the first run in a session and then every ten hours.
-#' This behavior can be disabled by setting the `check_blimp_update` option to FALSE using [`options`].
+#' This behavior can be disabled by setting the `check_blimp_update` option to FALSE using [`options`],
+#' or by setting the `R_BLIMP_NO_UPDATE_CHECK` environment variable to `"true"`.
 #' This check is not performed if R is not being run with an interactive session. See [`interactive`] for more information.
 #' @examplesIf has_blimp()
 #' # Run blimp script
@@ -24,8 +25,8 @@ rblimp_source <- function(
         output = TRUE,
         nopowershell = FALSE) {
 
-    # Check blimp update
-    if (getOption("check_blimp_update", default = TRUE)) {
+    # Check blimp update (gated by option AND env var)
+    if (update_check_enabled()) {
         if (check_blimp_update()) throw_error(c(
             "Updating Blimp",
             i = "Rerun previous model once updated."